requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
lxml==4.9.3
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Databricks & Spark
pyspark==3.5.0
databricks-sdk==0.15.0
delta-spark==2.4.0

# LLM Providers
openai==1.3.0
azure-identity==1.15.0
azure-keyvault-secrets==4.7.0

# Cloud Storage Connectors
boto3==1.29.0
google-cloud-storage==2.14.0
azure-storage-blob==12.19.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
jsonschema==4.20.0

# Testing (included for development)
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
mypy==1.7.0

# Development Tools
black==23.11.0
flake8==6.1.0
